+++ Fullstack.SAXS.Domain/Contracts/AreaParticleFactory.cs
@@ -5,13 +5,13 @@ namespace Fullstack.SAXS.Domain.Contracts
 {
     public abstract class AreaParticleFactory
     {
-        public abstract IEnumerable<Area> GetAreas(float areaSize, int number, float maxParticleSize);
+        public abstract IEnumerable<Area> GetAreas(double areaSize, int number, double maxParticleSize);
         public abstract IEnumerable<Particle> GetInfParticles(
-            float minSize, float maxSize, float sizeShape, float sizeScale,
-            float alphaRotation, float betaRotation, float gammaRotation,
-            float minX, float maxX,
-            float minY, float maxY,
-            float minZ, float maxZ
+            double minSize, double maxSize, double sizeShape, double sizeScale,
+            double alphaRotation, double betaRotation, double gammaRotation,
+            double minX, double maxX,
+            double minY, double maxY,
+            double minZ, double maxZ
         );
     }
 }
